+++ src/idc/imgaug/filter/__init__.py
@@ -13,9 +13,10 @@ from ._hsl_grayscale import HSLGrayScale
 from ._linear_contrast import LinearContrast
 from ._meta_sub_images import MetaSubImages
 from ._resize import Resize
+from ._roi_images import RegionOfInterestImages
 from ._rotate import Rotate
 from ._scale import Scale
 from ._sub_images import SubImages
 from ._sub_images_utils import (parse_regions, new_from_template, extract_regions, transfer_region, prune_annotations, region_filename,
                                 PLACEHOLDERS, REGION_SORTING, REGION_SORTING_XY, REGION_SORTING_YX, REGION_SORTING_NONE, DEFAULT_SUFFIX,
-                                Region, generate_regions, regions_to_string)
+                                Region, generate_regions, regions_to_string, locatedobject_to_region, locatedobject_to_xyxy)